UI improvements: square grid images, rating display, grocery list count, and scrape failure handling
- Add rating field to recipe schema and conditional display on public recipe page
- Handle scrape failure by opening recipe URL in browser
- Make grid images square with AspectRatio and show full titles (2 lines)
- Change all grids to 2-column layout across recipe page and profile pages
- Add grocery list item count badge next to cart icon on recipe page
- Update grocery list UI: X delete icon, move cart icon to submit button
- Change nav bar unselected colors to tertiary
- Change featured list text color to customColor3
- Fix addtocart unmounted widget error

+++ lib/backend/api_requests/api_calls.dart
@@ -123,7 +123,7 @@ class TestCall {
         response,
         r'''$.servings''',
       ));
-  static int? rating(dynamic response) => castToType<int>(getJsonField(
+  static double? rating(dynamic response) => castToType<double>(getJsonField(
         response,
         r'''$.rating''',
       ));
